requirements.txt
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
executing==2.0.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
gspread==6.1.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wsproto==1.2.0
